Revert "initial"
This reverts commit 3adcfb39c8f790be8a405fde5c8d1b9441e2c781.

+++ Content.Server/Station/Systems/StationSpawningSystem.cs
@@ -160,7 +160,7 @@ public sealed class StationSpawningSystem : SharedStationSpawningSystem
             if (profile.FlavorText != "" && _configurationManager.GetCVar(CCVars.FlavorText))
                 EnsureComp<DetailExaminableComponent>(entity.Value).Content = profile.FlavorText;
             var bank = EnsureComp<BankAccountComponent>(entity.Value);
-            bank.Balance = (int)profile.BankBalance;
+            bank.Balance = profile.BankBalance;
         }
 
         DoJobSpecials(job, entity.Value);
